Extract BetsFactory from NeoFoodClub god struct, dedupe cache invalidation
Follow-up to #171's deferred item: NeoFoodClub mixed raw-data facade,
cache management, and a full bet-construction factory in one 1045-line
file. Moves the bet-indices/ranking helpers, all make_* bet-construction
methods, and win-scoring methods into src/bets_factory.rs (no public API
change - all moved methods keep their exact names/signatures on
NeoFoodClub).

Also centralizes the cache-invalidation logic that clear_caches,
with_modifier, and set_bet_amount each independently duplicated across
3 places into one clear_ranking_caches() helper.
